Add benchmark scenario
- Add IncreasingNumOfStatesAndStackSymbols
- Minor fixes in scenarios

+++ src/benchmark/scenario/srs/CancellationAccuracy.cpp
@@ -10,7 +10,7 @@ void CancellationAccuracy::runSingleIteration()
         Tester{
             numOfTestsInSingleIteration,
             std::unique_ptr<generator::Generator>(new generator::srs::CancellationGenerator()),
-            parameters, rand()}
+            parameters, static_cast<uint16_t>(rand())}
             .run();
 }
