+++ src/benchmark/scenario/srs/CancellationIncreasingNumOfStates.cpp
@@ -19,7 +19,7 @@ void CancellationIncreasingNumOfStates::runSingleIteration()
     Tester{
         numOfTestsInSingleIteration,
         std::unique_ptr<generator::Generator>(new generator::srs::CancellationGenerator()),
-        parameters, rand()}
+        parameters, static_cast<uint16_t>(rand())}
         .run();
 }
 } // namespace benchmark::scenario
